[ores.qt.refdata,ores.qt.mktdata] Move CurrencyMarketTier UI to refdata plugin
currency_market_tier is refdata (served by ores.refdata.service, uses
refdata.v1.* NATS subjects, lives in ores.refdata.api/domain/) alongside
rounding_type and monetary_nature. Placing its Qt controller in MktdataPlugin
was wrong on two counts: the data is not market data, and the cross-plugin
signal boundary meant showMarketTiersRequested from CurrencyController was
never connected, leaving the market-tiers toolbar icon dead.

Move all 12 files (5 hpp, 5 cpp, 2 ui) from ores.qt/mktdata to ores.qt/refdata.
Wire CurrencyMarketTierController in RefdataPlugin alongside its siblings
(RoundingTypeController, MonetaryNatureController): on_login instantiation,
showMarketTiersRequested relay connect, Currency Codes menu entry, on_logout
reset. Remove from MktdataPlugin entirely.

Story-ID: 7F9C0F29-4268-4B62-A1BB-2153ECF0A858
Task-ID: E276A541-93EA-46EC-8CD8-D40803FEE587
