+++ projects/ores.qt/refdata/src/CurrencyMarketTierMdiWindow.cpp
