+++ projects/ores.qt/refdata/ui/CurrencyMarketTierHistoryDialog.ui
